fixing claude stuff
