+++ backend/app/repositories/data_type_coverage_repository.py
@@ -1,32 +1,16 @@
 from collections.abc import Iterable
 from datetime import datetime, timezone
-from typing import NamedTuple
 from uuid import UUID
 
-from sqlalchemy import event as sa_event
 from sqlalchemy import func, select
 from sqlalchemy.dialects.postgresql import insert
-from sqlalchemy.orm import Session
 
 from app.database import DbSession
 from app.models import DataTypeCoverage
+from app.schemas.data_type_coverage import CoverageSpan
 from app.schemas.sync_status import DataTypeKind
 
-# Spans queued on the session, merged and written by _flush_pending on commit.
-_PENDING_KEY = "data_type_coverage_pending"
-
 type _SpanKey = tuple[UUID, str, str, DataTypeKind]
-
-
-class CoverageSpan(NamedTuple):
-    """One data type's span within a single write, before it is merged into coverage."""
-
-    user_id: UUID
-    provider: str
-    data_type: str
-    kind: DataTypeKind
-    start: datetime
-    end: datetime
 
 
 class DataTypeCoverageRepository:
@@ -38,19 +22,38 @@ class DataTypeCoverageRepository:
     """
 
     def record(self, db_session: DbSession, spans: Iterable[CoverageSpan]) -> None:
-        """Queue spans to widen coverage when the caller's transaction commits.
+        """Widen coverage to include these spans, in the caller's transaction.
 
-        Merged on the way in, so the single-record write paths (one workout, one sleep
-        session) cost no round trip each and a whole import ends up as one statement.
-        Coverage is a min/max, so nothing is lost by deferring it.
+        Ranges only ever widen, so batches of one type arriving out of order still add up
+        to a single span. Runs in the caller's transaction and does not commit, so a
+        rollback takes the coverage with the rows it described.
         """
-        pending: dict[_SpanKey, tuple[datetime, datetime]] = db_session.info.setdefault(_PENDING_KEY, {})
-        for span in spans:
-            key = (span.user_id, span.provider, span.data_type, span.kind)
-            current = pending.get(key)
-            pending[key] = (
-                (span.start, span.end) if current is None else (min(current[0], span.start), max(current[1], span.end))
+        rows = [
+            {
+                "user_id": user_id,
+                "provider": provider,
+                "data_type": data_type,
+                "kind": kind,
+                "coverage_start": start,
+                "coverage_end": end,
+                "last_written_at": datetime.now(timezone.utc),
+            }
+            for (user_id, provider, data_type, kind), (start, end) in _merge(spans).items()
+        ]
+        if not rows:
+            return
+
+        stmt = insert(DataTypeCoverage).values(rows)
+        db_session.execute(
+            stmt.on_conflict_do_update(
+                index_elements=["user_id", "provider", "data_type"],
+                set_={
+                    "coverage_start": func.least(stmt.excluded.coverage_start, DataTypeCoverage.coverage_start),
+                    "coverage_end": func.greatest(stmt.excluded.coverage_end, DataTypeCoverage.coverage_end),
+                    "last_written_at": stmt.excluded.last_written_at,
+                },
             )
+        )
 
     def list_for_user(
         self,
@@ -75,44 +78,16 @@ class DataTypeCoverageRepository:
         return db_session.get(DataTypeCoverage, (user_id, provider, data_type))
 
 
-@sa_event.listens_for(Session, "before_commit")
-def _flush_pending(session: Session) -> None:
-    """Write the queued spans in the committing transaction, so they cannot outlive it."""
-    pending: dict[_SpanKey, tuple[datetime, datetime]] = session.info.pop(_PENDING_KEY, {})
-    if not pending:
-        return
-
-    written_at = datetime.now(timezone.utc)
-    stmt = insert(DataTypeCoverage).values(
-        [
-            {
-                "user_id": user_id,
-                "provider": provider,
-                "data_type": data_type,
-                "kind": kind,
-                "coverage_start": start,
-                "coverage_end": end,
-                "last_written_at": written_at,
-            }
-            for (user_id, provider, data_type, kind), (start, end) in pending.items()
-        ]
-    )
-    session.execute(
-        stmt.on_conflict_do_update(
-            index_elements=["user_id", "provider", "data_type"],
-            set_={
-                "coverage_start": func.least(stmt.excluded.coverage_start, DataTypeCoverage.coverage_start),
-                "coverage_end": func.greatest(stmt.excluded.coverage_end, DataTypeCoverage.coverage_end),
-                "last_written_at": stmt.excluded.last_written_at,
-            },
+def _merge(spans: Iterable[CoverageSpan]) -> dict[_SpanKey, tuple[datetime, datetime]]:
+    """Collapse spans to one per key: Postgres cannot upsert the same key twice per statement."""
+    merged: dict[_SpanKey, tuple[datetime, datetime]] = {}
+    for span in spans:
+        key = (span.user_id, span.provider, span.data_type, span.kind)
+        current = merged.get(key)
+        merged[key] = (
+            (span.start, span.end) if current is None else (min(current[0], span.start), max(current[1], span.end))
         )
-    )
-
-
-@sa_event.listens_for(Session, "after_rollback")
-def _discard_pending(session: Session) -> None:
-    """Drop spans whose rows were rolled back. Savepoint rollbacks do not reach here."""
-    session.info.pop(_PENDING_KEY, None)
+    return merged
 
 
 data_type_coverage_repository = DataTypeCoverageRepository()
