+++ backend/app/repositories/event_record_repository.py
@@ -28,8 +28,9 @@ from sqlalchemy.orm import Query, selectinload
 from app.database import DbSession
 from app.models import DataPointSeries, DataSource, EventRecord, SleepDetails, WorkoutDetails
 from app.repositories.data_source_repository import DataSourceRepository
-from app.repositories.data_type_coverage_repository import CoverageSpan, data_type_coverage_repository
+from app.repositories.data_type_coverage_repository import data_type_coverage_repository
 from app.repositories.repositories import CrudRepository
+from app.schemas.data_type_coverage import CoverageSpan
 from app.schemas.enums import ProviderName, SeriesType, get_series_type_id
 from app.schemas.model_crud.activities import (
     EventRecordCreate,
